# Skillbox/Module_18_OpenAPI/HW_laptop/swagger.yaml
definitions:
  Author:
    properties:
      first_name:
        type: string
      id:
        readOnly: true
        type: integer
      last_name:
        type: string
      middle_name:
        type: string
    required:
      - first_name
      - last_name
    type: object
  Book:
    properties:
      author: {}
      id:
        readOnly: true
        type: integer
      title:
        type: string
    required:
      - author
      - title
    type: object
info:
  title: Books API
  version: 1.0.0
paths:
  /api/books:
    get:
      responses:
        '200':
          description: Success get all books
          schema:
            items:
              $ref: '#/definitions/Book'
            type: array
      summary: Get all books
      tags:
        - books
    post:
      parameters:
        - in: body
          name: new_book
          required: true
          schema:
            $ref: '#/definitions/Book'
      responses:
        '201':
          description: Book added
          schema:
            $ref: '#/definitions/Book'
            type: object
      summary: Add book
      tags:
        - books
  /api/books/{id}:
    delete:
      responses:
        '204':
          description: Success delete book by id
          schema: {}
      summary: Delete book by id
      tags:
        - books/id
    get:
      responses:
        '200':
          description: Success get book by id
          schema:
            $ref: '#/definitions/Book'
            type: object
      summary: Get book by id
      tags:
        - books/id
    patch:
      parameters:
        - in: body
          name: patch_data_book
          schema:
            $ref: '#/definitions/Book'
            type: object
      responses:
        '200':
          description: Success patch book by id
          schema:
            $ref: '#/definitions/Book'
            type: object
      summary: Patch book by id
      tags:
        - books/id
    put:
      parameters:
        - in: body
          name: update_data_book
          schema:
            $ref: '#/definitions/Book'
            type: object
      responses:
        '200':
          description: Success put book by id
          schema:
            $ref: '#/definitions/Book'
            type: object
      summary: Update book by id
      tags:
        - books/id
swagger: 2.0.0